+++ tests/test_events.py
@@ -20,6 +20,7 @@ from watchdog.events import (
     FileModifiedEvent,
     FileCreatedEvent,
     FileClosedEvent,
+    FileOpenedEvent,
     DirDeletedEvent,
     DirModifiedEvent,
     DirCreatedEvent,
@@ -31,6 +32,7 @@ from watchdog.events import (
     EVENT_TYPE_DELETED,
     EVENT_TYPE_MOVED,
     EVENT_TYPE_CLOSED,
+    EVENT_TYPE_OPENED,
 )
 
 path_1 = '/path/xyz'
@@ -92,6 +94,14 @@ def test_file_closed_event():
     assert not event.is_synthetic
 
 
+def test_file_opened_event():
+    event = FileOpenedEvent(path_1)
+    assert path_1 == event.src_path
+    assert EVENT_TYPE_OPENED == event.event_type
+    assert not event.is_directory
+    assert not event.is_synthetic
+
+
 def test_dir_deleted_event():
     event = DirDeletedEvent(path_1)
     assert path_1 == event.src_path
@@ -122,6 +132,7 @@ def test_file_system_event_handler_dispatch():
     dir_cre_event = DirCreatedEvent('/path/blah.py')
     file_cre_event = FileCreatedEvent('/path/blah.txt')
     file_cls_event = FileClosedEvent('/path/blah.txt')
+    file_opened_event = FileOpenedEvent('/path/blah.txt')
     dir_mod_event = DirModifiedEvent('/path/blah.py')
     file_mod_event = FileModifiedEvent('/path/blah.txt')
     dir_mov_event = DirMovedEvent('/path/blah.py', '/path/blah')
@@ -137,6 +148,7 @@ def test_file_system_event_handler_dispatch():
         file_cre_event,
         file_mov_event,
         file_cls_event,
+        file_opened_event,
     ]
 
     class TestableEventHandler(FileSystemEventHandler):
@@ -158,6 +170,9 @@ def test_file_system_event_handler_dispatch():
 
         def on_closed(self, event):
             assert event.event_type == EVENT_TYPE_CLOSED
+
+        def on_opened(self, event):
+            assert event.event_type == EVENT_TYPE_OPENED
 
     handler = TestableEventHandler()