+++ changelog.rst
@@ -8,8 +8,9 @@ Changelog
 
 2023-xx-xx • `full history <https://github.com/gorakhargosh/watchdog/compare/v2.2.1...HEAD>`__
 
+- [inotify] Add support for ``IN_OPEN`` events: a ``FileOpenedEvent`` event will be fired. (`#941 <https://github.com/gorakhargosh/watchdog/pull/941>`__)
 - [watchmedo] Add optional event debouncing for ``auto-restart``, only restarting once if many events happen in quick succession (`#940 <https://github.com/gorakhargosh/watchdog/pull/940>`__)
-- Thanks to our beloved contributors: @BoboTiG, @taleinat
+- Thanks to our beloved contributors: @BoboTiG, @dstaple, @taleinat
 
 2.2.1
 ~~~~~